requirements.txt
aniso8601==9.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask-RESTful==0.3.9
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pip==23.0.1
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
setuptools==65.5.1
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.3
wheel==0.38.4
